Lots of changes, esp. wrt. i3
Added i3 configuration to repository, which starts Dropbox and a script
I've written which allows us to have different wallpapers for different
i3 workspaces.

Further to this, I've added a selection of wallpapers to the wallpapers
directory, all from Simple Desktops (http://simpledesktops.com). These
are in the repository, but are in no way my creation.

Added the vim-speeddating plugin, as vim-orgmode seemed to want it. Also
added the spelling dictionary add-file to the repository, though it's
not really got anything in it. Updated SurvivalGuide to reflect
disappearance of Molly.

Modified the install script to finally work as intended and only update
symlinks which don't already exist.